Move code from it's own class into the DALLE class. Few text tweaks
